portfolio POST functionality (WIP)
cleaning and reorganizing test infra code to allow for expansion of test utilities
