+++ src/main/go/inttest/allocation_plan_integration_test.go
@@ -1,7 +1,7 @@
 package inttest
 
 import (
-	"github.com/benizzio/open-asset-allocator/inttest/util"
+	inttestinfra "github.com/benizzio/open-asset-allocator/inttest/infra"
 	"github.com/stretchr/testify/assert"
 	"io"
 	"net/http"
@@ -10,7 +10,7 @@ import (
 
 func TestGetAllocationPlans(t *testing.T) {
 
-	response, err := http.Get(util.TestAPIURLprefix + "/portfolio/1/allocation-plan")
+	response, err := http.Get(inttestinfra.TestAPIURLprefix + "/portfolio/1/allocation-plan")
 	assert.NoError(t, err)
 
 	assert.Equal(t, http.StatusOK, response.StatusCode)
